why chid failed

+++ kernel/proc.h
@@ -93,7 +93,7 @@ struct proc {
   int xstate;                  // Exit status to be returned to parent's wait
   int pid;                     // Process ID
 
-  int trace_id;
+  uint64 trace_id;
   // wait_lock must be held when using this:
   struct proc *parent;         // Parent process
